fix: rename js files
